move initialMessage module to optimize build
